hukum_ai/backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
chromadb
python-dotenv
passlib[bcrypt]
pydantic
email-validator